More cleanups, put unicode in the right place
