+++ crates/bridge_icu/Cargo.toml
@@ -20,6 +20,9 @@ license = "MIT"
 edition = "2021"
 links = "icuuc"
 
+[dependencies]
+libc = "0.2"
+
 [build-dependencies]
 tectonic_dep_support = { path = "../dep_support", version = "0.0.0-dev.0" }
